requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==21.5b1
click==8.0.0
distlib==0.3.1
Django==3.1.8
django-appconf==1.0.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.1.0
django-redis==4.12.1
django-select2==7.7.1
djangorestframework==3.12.4
filelock==3.0.12
isort==5.6.4
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumbers==8.12.13
Pillow==8.1.1
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
regex==2021.4.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.2.0
